Refactor CSS for header and home page, Update translations.js and templates, Add redirects on home page.
The CSS for the header and home pages have been refactored to improve the style. The translations.js file has been updated to provide the translations for the various languages supported. The templates have also been updated to support these changes. Additionally, redirects have been added to the home page, redirecting users to the name.html page when they click on the "play" button.