feat(HomeController): add endpoints for hosting and joining a race
This commit adds endpoints for creating and joining an instance of Race. The
'/host' endpoint creates a new race given a Player object and returns the newly
created race. The '/join/{raceID}' endpoint unpacks a JoinRequest containing a
Player object and a pre-existing race ID. If successful, the endpoint adds the
Player to the corresponding race. Additionally, new HTML templates for pages
related to race management are added: lobby, host-join, join, host, and race.
One existing HTML template for the game's tutorial, watQrace.html, is removed.